requirements.txt
google-auth==1.6.3
google-auth-oauthlib==0.4.1
google-auth-httplib2==0.0.3
google-api-python-client==1.7.11
pytz==2021.1
django>=3.2,<4.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
psycopg2-binary==2.9.3
snowflake-id
pybase62

